+++ src/ambition/version_details.cc
@@ -28,6 +28,7 @@ std::vector<VersionUpdates> versionDetails = {
         .heading = _(HEADING_FIXED),
         .items = {
           _("Fixed target resistance being shown incorrectly when you have workers in the independent town."),
+          _("Fixed the Harbour production count not being placed properly on the button."),
         },
       },
       {